VOAC-7: Merged council and business confirmation pages onto a single page
