+++ app/uk/gov/hmrc/valuationofficeagencycontactfrontend/views/businessRatesSubcategory.scala.html
@@ -30,7 +30,7 @@
     appConfig = appConfig,
     bodyClasses = None) {
 
-    @components.back_link(uk.gov.hmrc.valuationofficeagencycontactfrontend.controllers.routes.EnquiryCategoryController.onPageLoad(NormalMode).url)
+    @components.back_link(uk.gov.hmrc.valuationofficeagencycontactfrontend.controllers.routes.ContactDetailsController.onPageLoad(NormalMode).url)
 
     @components.heading("businessRatesSubcategory.heading")
