+++ app/uk/gov/hmrc/valuationofficeagencycontactfrontend/views/confirmation.scala.html
@@ -19,12 +19,12 @@
 @import uk.gov.hmrc.valuationofficeagencycontactfrontend.utils.AddressFormatters._
 @import uk.gov.hmrc.valuationofficeagencycontactfrontend.utils.ContactFormatter._
 
-@(appConfig: FrontendAppConfig, businessContact: Contact, date: String)(implicit request: Request[_], messages: Messages)
+@(appConfig: FrontendAppConfig, contact: Contact, date: String)(implicit request: Request[_], messages: Messages)
 
 @main_template(
     title = messages("confirmation.title"),
     appConfig = appConfig,
-    bodyClasses = Some("full-width")) {
+    bodyClasses = Some("")) {
     <br />
     <br />
     <div class="transaction-banner--complete">
@@ -37,22 +37,22 @@
     <ul class="tabular-data tabular-data--list text">
         <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
             <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.enquiryType")</div>
-            <div class="tabular-data__data tabular-data__data--8-12">@messages(s"enquiryCategory.${businessContact.enquiryCategory}") <br />
-                @messages(s"businessRatesSubcategory.${businessContact.subEnquiryCategory}")</div>
+            <div class="tabular-data__data tabular-data__data--8-12">@messages(s"enquiryCategory.${contact.enquiryCategory}") <br />
+                @messages(s"councilTaxSubcategory.${contact.subEnquiryCategory}")</div>
         </li>
         <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
             <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.yourDetails")</div>
             <div class="tabular-data__data tabular-data__data--8-12">
-            @Html(formattedConfirmedContactDetails(businessContact.contact, "<br/>"))
+            @Html(formattedConfirmedContactDetails(contact.contact, "<br/>"))
             </div>
         </li>
         <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
             <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.propertyAddress")</div>
-            <div class="tabular-data__data tabular-data__data--8-12">@Html(formattedPropertyAddress(businessContact.propertyAddress, "<br/>"))</div>
+            <div class="tabular-data__data tabular-data__data--8-12">@Html(formattedPropertyAddress(contact.propertyAddress, "<br/>"))</div>
         </li>
         <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
             <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.yourMessage")</div>
-            <div class="tabular-data__data tabular-data__data--8-12 word-break">@businessContact.message</div>
+            <div class="tabular-data__data tabular-data__data--8-12  word-break">@contact.message</div>
         </li>
         <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
             <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.date")</div>
@@ -66,4 +66,5 @@
     <p><a href="javascript:window.print()" class="button print-hidden js-visible" id="print-button">@messages("site.print.button")</a></p>
     <br />
     <p><a href="http://www.gov.uk" id="backToGovUk">@messages("site.govuk")</a></p>
+
 }
