--- app/uk/gov/hmrc/valuationofficeagencycontactfrontend/views/confirmationCouncilTax.scala.html
@@ -1,70 +1,0 @@
-@*
- * Copyright 2017 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *@
-
-@import uk.gov.hmrc.valuationofficeagencycontactfrontend.FrontendAppConfig
-@import uk.gov.hmrc.valuationofficeagencycontactfrontend.models.Contact
-@import uk.gov.hmrc.valuationofficeagencycontactfrontend.utils.AddressFormatters._
-@import uk.gov.hmrc.valuationofficeagencycontactfrontend.utils.ContactFormatter._
-
-@(appConfig: FrontendAppConfig, councilContact: Contact, date: String)(implicit request: Request[_], messages: Messages)
-
-@main_template(
-    title = messages("confirmation.title"),
-    appConfig = appConfig,
-    bodyClasses = Some("")) {
-    <br />
-    <br />
-    <div class="transaction-banner--complete">
-        <h1 class="transaction-banner__heading">@messages("confirmation.heading")</h1>
-        <p>@messages("confirmation.para1")</p>
-    </div>
-
-    <h2 class="heading-medium">@messages("confirmation.enquirySummary")</h2>
-
-    <ul class="tabular-data tabular-data--list text">
-        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
-            <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.enquiryType")</div>
-            <div class="tabular-data__data tabular-data__data--8-12">@messages(s"enquiryCategory.${councilContact.enquiryCategory}") <br />
-                @messages(s"councilTaxSubcategory.${councilContact.subEnquiryCategory}")</div>
-        </li>
-        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
-            <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.yourDetails")</div>
-            <div class="tabular-data__data tabular-data__data--8-12">
-            @Html(formattedConfirmedContactDetails(councilContact.contact, "<br/>"))
-            </div>
-        </li>
-        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
-            <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.propertyAddress")</div>
-            <div class="tabular-data__data tabular-data__data--8-12">@Html(formattedPropertyAddress(councilContact.propertyAddress, "<br/>"))</div>
-        </li>
-        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
-            <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.yourMessage")</div>
-            <div class="tabular-data__data tabular-data__data--8-12  word-break">@councilContact.message</div>
-        </li>
-        <li class="tabular-data__entry tabular-data__entry--grouped divider--bottom">
-            <div class="tabular-data__data tabular-data__data--4-12">@messages("confirmation.section.date")</div>
-            <div class="tabular-data__data tabular-data__data--8-12  word-break">@date</div>
-        </li>
-    </ul>
-    <br />
-    <h2 class="heading-medium">@messages("confirmation.whatHappensnext")</h2>
-    <p class="text">@messages("confirmation.para2")</p>
-
-    <p><a href="javascript:window.print()" class="button print-hidden js-visible" id="print-button">@messages("site.print.button")</a></p>
-    <br />
-    <p><a href="http://www.gov.uk" id="backToGovUk">@messages("site.govuk")</a></p>
-
-}
