+++ test/uk/gov/hmrc/valuationofficeagencycontactfrontend/views/BusinessRatesSubcategoryViewSpec.scala
@@ -69,12 +69,12 @@ class BusinessRatesSubcategoryViewSpec extends ViewBehaviours {
         assert(continueButton == messages("site.continue"))
       }
 
-      "has a link marked with site.back leading to the Enquiry Category Page" in {
+      "has a link marked with site.back leading to the Contact Details Page" in {
         val doc = asDocument(createView())
         val backlinkText = doc.select("a[class=link-back]").text()
         backlinkText mustBe messages("site.back")
         val backlinkUrl = doc.select("a[class=link-back]").attr("href")
-        backlinkUrl mustBe uk.gov.hmrc.valuationofficeagencycontactfrontend.controllers.routes.EnquiryCategoryController.onPageLoad(NormalMode).url
+        backlinkUrl mustBe uk.gov.hmrc.valuationofficeagencycontactfrontend.controllers.routes.ContactDetailsController.onPageLoad(NormalMode).url
       }
     }
